fix: move sort order
