+++ src/config.ts
@@ -1,7 +1,7 @@
 import { FuzzySearchConfig } from './fuzzy-searchers/fuzzy-search-config.js';
 import { NormalizerConfig } from './normalization/normalizer-config.js';
 import { SearcherType } from './interfaces/searcher-type.js';
-import { SortOrder } from './sort-order.js';
+import { SortOrder } from './interfaces/sort-order.js';
 import { SubstringSearchConfig } from './suffix-array-searchers/substring-search-config.js';
 
 /**
@@ -24,7 +24,7 @@ export class Config {
     public normalizerConfig: NormalizerConfig,
     public fuzzySearchConfig?: FuzzySearchConfig,
     public substringSearchConfig?: SubstringSearchConfig
-  ) {}
+  ) { }
 
   /**
    * Creates an opinionated default configuration.
