+++ src/suffix-array-searchers/index.ts
@@ -3,4 +3,3 @@ export { SuffixArray } from './suffix-array.js';
 export { SuffixArraySearcher } from './suffix-array-searcher.js';
 export { SubstringSearchConfig } from './substring-search-config.js';
 export { PrefixSearcher as PrefixSearcherImpl } from './prefix-searcher.js';
-export { SortOrder } from '../sort-order.js';
